swapi_api/requirements.txt
aniso8601==9.0.1
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==4.2.15
django-filter==24.3
djangorestframework==3.15.2
exceptiongroup==1.2.2
graphene==3.3
graphene-django==3.2.2
graphql-core==3.2.3
graphql-relay==3.2.0
gunicorn==23.0.0
idna==3.7
importlib_metadata==8.2.0
iniconfig==2.0.0
Markdown==3.7
numpy==2.0.1
packaging==24.1
pluggy==1.5.0
promise==2.3
psycopg2-binary==2.9.9
pytest==8.3.2
pytest-django==4.8.0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.12.2
urllib3==2.2.2
zipp==3.20.0
